+++ app/utils/defaultSettings.js
@@ -30,6 +30,8 @@ module.exports = {
   breakPostponesLimit: 1,
   breakPostponableDurationPercent: 30,
   mainColor: '#478484',
+  transparentMode: false,
+  opacity: 0.9,
   audio: 'crystal-glass',
   volume: 1,
   fullscreen: false,